backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend to FastAPI Template"
authors = ["Eric <erickh.li@mail.utoronto.ca>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.73.0"
pytest = "^6.2.5"
SQLAlchemy = "^1.4.31"
requests = "^2.27.1"
SQLAlchemy-Utils = "^0.38.2"
pydantic = "^1.9.0"
uvicorn = "^0.17.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.3"
strawberry-graphql = "^0.95.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
